Initial Python SmartGadget for Linux
The SmartGadget module supports SHTC1 and SHT3x Gadgets. However, the
subscription of live values currently on works on SHTC1.

Works best with BlueZ >= 4.41 but, unfortunately, the state of BLE
on Linux is pretty crappy and may require restarts and special
rights.

This only works on Linux (PyGatt module works directly on the command
line utilities `hcitool' and `gatttool').

See __main__ section of SmartGadget.py for details, tips and sample
usage.

Change-Id: I1719270c724b9916cf2400c42857c94ff8084f2c